refactor: 最小化宏重构基本完毕
