+++ src/policy/usage/usage_top1/lolm/common.rs
@@ -1,2 +1,2 @@
-use super::super::macro_common::top1_common;
-top1_common!(Only6);
+// use super::super::macro_common::top1_common;
+// top1_common!(Only6);
