+++ src/policy/usage/usage_top1/unnamed/common.rs
@@ -1,2 +1,2 @@
-use super::super::macro_common::top1_common;
-top1_common!(Only7);
+// use super::super::macro_common::top1_common;
+// top1_common!(Only7);
