+++ src/policy/usage/usage_top1/unnamed/policy_top1.rs
@@ -1,7 +1,7 @@
-use super::common::Policy;
+use crate::policy::usage::usage_top1::macro_common::Policy;
 use crate::policy::{
     pkg_cfg::StartArgs,
-    usage::{check_some, get_thread_tids, usage_top1::top1_policy, UNNAME_TIDS},
+    usage::{check_some, get_thread_tids, usage_top1::top1_macro_init, UNNAME_TIDS},
 };
 use likely_stable::{likely, unlikely};
 #[cfg(debug_assertions)]
@@ -14,4 +14,4 @@ const ONLY7: [&[u8]; 0] = [];
 const MIDDLE: [&[u8]; 1] = [b"RenderThread"];
 const BACKEND: [&[u8]; 0] = [];
 
-top1_policy!(b"Thread-");
+top1_macro_init!(b"Thread-");
